migrate server/index file + parse-oembed file to typescript
