+++ packages/slice-machine/package.json
@@ -87,6 +87,7 @@
     "@theme-ui/color": "0.8.4",
     "@theme-ui/prism": "0.8.4",
     "@types/ejs": "^3.0.6",
+    "@types/express": "^4.17.13",
     "@types/faker": "^5.5.5",
     "@types/glob": "^7.1.3",
     "@types/jest": "^26.0.23",
