+++ packages/slice-machine/server/src/api/parse-oembed.ts
@@ -1,6 +1,6 @@
 const extract = require("oembed-parser").extract;
 
-export default async function handler(url) {
+export default async function handler(url: string) {
   try {
     const oembed = await extract(url);
     return { oembed, err: null };
